Rework match rules
Test bed is showing what works and what doesn't. This change
switches back to a variation based on an earlier approach
that compacts subscriptions into a 64 bit value that can
be intersected with other subscriptions to form a useful
match filter. This along with a population count check is
much more effective than the whitening approach.