Identify publications by UUID
Changes how publications are identified. Now they are identified by a UUID. The same
publication can be updated with a new payload and this publication identity is
preserved but new values get a serial number. This enables data streams. The UUIDs
are also used for retained messages and for identifying stale publications.