Reworking sub forwarding
Subscription forwarding is tricky. Interests forwarded downstream need to exclude
the subscriptions of the downstream tree otherwise the downstream tree will see itself
as having interests it doesn't have subscribers for. This commit replaces the
simple bitvectors in the node with counting vectors that make is easier to exclude
specific interests for each remote node.