Rework how scope-id is set
Implementation assumed the interface index was the same as the index of interface
in the interface list. It is not, so if_indexfromname() must be called to get the
correct index for the interface. Also, the scope-id only needs to be set for
link local addresses. Other minor change is to cancel writes from the shutdown
callback because the node lock cannot be held when the data received callback
is made.