Encode bit vectors are cbor arrays
Normalize the encoding of bit vectors as a CBOR array with
three elements - flags, bit vector length, and the possibly
compressed bit vector. Also in this commit a new CBOR function
CBOR_Skip() that will be needed for the COSE integration.
Test cases for CBOR_Skip added to cbortest.c