Integrate COSE encryption
This commit implements COSE encryption for publications and acknowledgments.
The change is large because the serialization had to be reworked to organize
the message contents into separate blocks for data that is unprotected and
changes hop-by-hop such as the originating port and current ttl, protected
content which is authenticated but send in the clear such a pulbication ids,
and bloom filters, and content which is encrypted - topics lists and payloads.
API changes were also needed to expose the functionality needed for providing
encryption keys.

A change that was not included in this commit but will follow later will
clean up buffer management by creating a clear distinction betweem buffers
that are being written and buffers that are being read.