Separate DPS_Buffer into separate rx/tx buffer types
The dual use of DPS_Buffer for encoding and decoding make the
code hard to understand. This commit replaces DPS_Buffer
by two types DPS_RxBuffer which is nominally for decoding
received data and DPS_TxBuffer for encoding data to be sent.
The CBOR decoding functions take an Rx buffer, the encoding
functions take a Tx buffer. Functions are provided for converting
a Tx buffer into an Rx buffer and visa versa.