Refresh SWIG generated python binding
The python binding had fallen behind some recent API
changes and the sample scripts were no longer working.
This commit refreshes the binding and the scripts.
There are also a couple of minor changes (added const)
to the public API and two new functions for accessing
the DPS_Node from an subscription or publication.