Adds support for publication-specific key ids
The wire protocol allows unique kids (key identifiers) per message.
Previously the kid was set once when the node is created. This change
allows a kid to be passed when initializing a publication allowing
the use of multiple encryption keys. Acknowledgements were implicitly
encrypted using the same key as the publication but with this change
it is now explicit that ACKs use the same kid (and therefore key) as
the PUB being acknowledged.