Build script changes
Adds support for specifying location of libuv include and
libraries on Linux (was already supported on Windows)

Adds a build option for selectively suppressing building the
python and nodejs bindings. If SWIG is not installed the
build fails with a cryptic error message - SWIG is not
needed if "bindings=none" is set on the scons command line.

Adds a compile time check of the libuv version.