Add wait option to examples
Adds a command line option to wait for some time between
starting local node and linking to other nodes. This
allows for more deterministic behavior when setting up
a network of nodes by ensuring that all nodes are listening
on an open socket before any node starts to send data,
Removes the -b option from publisher.c and replaces it with
a short delay to ensure the publication is sent before
the link is dropped.