Enable loop detection
This change adds support for detecting loops in the mesh.
A loop is detected when the mesh id in a subscription
received from a remote node is the same as the mesh id
received from another node. The link to the remote node
identified as causing the loop is "muted" by sending the
remote node an empty interests vector.