Monitor muted links
Adds support for monitoring links that have been muted
to mitigate mesh loops so the muted link can be unmuted
if loop becomes broken. Monitoring is achieved by having
the node that muted the link subscribe to a unique topic
string and periodically send a publication to for that
topic string via the muted link. If the loop is intact
the sender will receive the publication. If the
publication is not received the muted link is unmuted
to restore mesh connectivity.