New debug option
This changes add a new debug flag that can be used to force debug
output for a specific module either at compile time or from within
a debugger. To enable debug output at compile time add this line to
the start of the module:

DPS_DEBUG_CONTROL(DPS_DEBUG_FORCE)

From within a debugger set the static variable  __DPS_DebugControl=2
for the module to enable.