Add DPS_KeyStore and a memory-only implementation
With the upcoming support for DTLS, we will have other types of keys
besides the ones used with COSE for encrypting the content. To
accomodate that, this patch replaces the key request callback with an
"key store" that can grow to handle other cases in the future.

The DPS_KeyStore is the low-level interface. This patch also adds
DPS_MemoryKeyStore, which "implements" a KeyStore that also has an
array of keys and a function to add keys.

The creates a memory key store, set the keys, then use it to create
the node. This will lead to less boilerplate when we add support for
other types of keys. Advanced users can still fallback to the
low-level interface provided by DPS_KeyStore.

In the future, it is a good idea to add a DPS_FileKeyStore, that get
the keys from a file instead of expect them to be set by the program
code.

The C samples were changed to use DPS_MemoryKeyStore.

The Python bindings were changed to expose DPS_MemoryKeyStore. Most of
the heavy-lifting is now done by SWIG. The Python samples were changed
to use DPS_MemoryKeyStore.

The Node/JS bindings accepted a callback but were not using it in the
binding code, so the functionality was not working. I've simply
removed the callback in the Python samples. As proven by the Python
code it should be straighforward to add support for DPS_MemoryKeyStore
once needed.