Rework of link recovery
This change ensures that the nodes on both sides of the muted
link are aware that the link is muted. The passive side of the
link is responsible for link monitoring and for unmuting the
link. When a link is unmuted a new mesh id lexically less than
any other known mesh id must be generated.