Improve graph layout
The graphviz "neato" engine does the best job of displaying
DPS node graphs but neato doesn't support clusters so the
label text gets lost. This change keeps the clusters but
rather than labelling the clusters puts the labels in a node
with invisible edges to a node in each subgraph it labels.
Neato is now the default (and preferred engine)

This change also adds coloring to the nodes to make it easier
to identify the before and after graphs when nodes are killed.
The before graph is green, the after graph is blue. Nodes that
are going to be killed are a lighter color (previously yellow)