Partially reverts change that broke loop detection
An earlier change to optimize/cleanup the loop detection code changed the behavior and broke loop detecion. This is a partial backout of that earlier change. Also in this change is a new test case for stress testing loop detection. The test case also exposes some problems with calls to non-thread-safe libuv functions to be addressed in an upcoming commit.