Clarify how sync flags are mapped
The sync flags in the protocol did not map cleanly to the sync flags in the remote struct leading to some confusion about usage. This change separates out the request (REQ) flags from the indication (IND) flags. This change uncovered a bug in the lost subscription recovery code where a sync request could be satisfied by a delta response. The SYNC_REQ flag is now only cleared when a SYNC_IND is received.  Subscription sequence numbers are now only incremented when the subscription has changed so the increment was moved into UpdateOutboundInterests()