Fix count vector intersection corner case
Function to compute the intersection from a count vector was returning a bit vector with all bits set if the input count vector was empty. Intersection is defined as each counter having the same value as the total number of entries but if the total number of entries is zero then the intersection is also zero.