Rearrange subscription header
This is the first of several commits that will eventually implement subscription acknowledgements. This commmit renames subscription sequence numbers as "revision numbers" and swaps the locations of the revision number and mesh id. The lost subscription simulation code has been move into the receive path rather than the send path. Removed the DumpFlags() debug print because it is not really useful.