Implement subscription acknowledgements
This change adds a new subscription ack message to the protocol that enables reliable subscription delivery. Now that subscriptions are reliable the previous synchronization mechanism is redundant and has been removed. Other than the initial subscription (and the first subscription after unmuting) all subscription interests are sent as deltas. This simplifies much of the logic in DPS_DecodeSubscription. Random message loss can be tested by compiling with SIMULATE_PACKET_LOSS set to a non-zero value. SIMULATE_PACKET_LOSS=100 causes one out of 100 subscriptions and subscription acks to be dropped.