No need to set mesh id until there is a subscription
If a node has no subscriptions it does not need a real mesh id. When a node is created the mesh id is set to the max mesh id value and only set to a "real" value when a subscription is registered. This gives a more balanced tree when running various test cases. This change is probably largely cosmetic because in real use cases it will be rare to have node with multiple network connections that does not also have subscriptions.

The make_mesh test case has been modified to highlight the nodes that have subscriptions. Also fixes some compiler warnings.