Fix stale criteria for publications
A publication is stale if it is not newer. The same sequence number
means a stale publication.

This fixes the case of a publisher that sends and receives
multicast. The code would proceed and end up removing the publication
from the list, a subsequent publish with the same publication would
fail.