Improve the check for stale publications
Check for stale publication before creating the data structure and
adding it to the list of publications. This avoids having to deal with
removing it in the failure case. Now this case correctly returns
DPS_ERR_STALE.

With this change we also don't perform a second stale check for
retained publications, the first check is sufficient.