Fixes for multicast publications
History was not being updated for multicast publications so stale publications were not being correctly identified and ignored. When a publication has a zero timeout the retained flag must to be cleared and also the expired flag in the case that a previous retained publication was expired.

Adds a simple python test case that exercises some new multicast usage scenarios.