Allow CBOR-encode double <-> float coersion
If a CBOR encoded double fits in the range of a float we permit is to be decoded into a float (with potential loss of
precision). Similarly CBOR encoded floats can now be decoded as doubles. This behavior is more consistent with the way integer
decoding is handled. A new error status DPS_ERR_RANGE flags decode errors for all numeric types where the CBOR encode
was correct but the decoded value cannot be represented within the range of the target type.