Add sanity check for CBOR_Peek
A common use for CBOR_Peek is to get the length for strings, arrays, maps, etc
so that the caller knows how much space to allocate for subsequent decoding.
An attacker could generate a packet with a large length and cause the application
to attempt to do a large memory allocation. This change adds a santity check that
ensures that in the worst case the length would not cause an allocation larger than
the remaining data in the CBOR packet.

Signed-off-by: Greg Burns <gregory.burns@intel.com>