Convert CBOR int encodings to float or double
CBOR encodings for integers use the smallest possible encoding
depending on the value. On decode the value is automatically converted
provided the converted value is in the required range. This change allows
a similar conversion from integer encoded CBOR to float or double provided
the conversion does not reuslt in a loss of precision. If precision is lost
the value is still converted but a new error status DPS_ERR_LOST_PRECISION
is returned allowing the application to decide how the handle this case.

Signed-off-by: Greg Burns <gregory.burns@intel.com>