Fix node discovery with dps middleware in ROS2
Currently node discovery does not work with dps middleware in ROS2.

The issue is as follows: When a subscriber comes online it sends out a
multicast discovery message. All the subscriber topics (bloom filter
outputs) are merged in a union and sent out as an "interest" message.

When the multicast message is received by the publishing node, it checks
whether any topic is in the set of the received interest. If a match is
found a unicast connection is created and the published message is sent
to the subscriber.

However at the moment, the publisher's bloom filter output takes into
account wildcards whereas they are not considered in the subscriber.
Thus the  publishers bloom filter output will not be in the set of the
subscriber's interests.  Hence, the unicast connection is not created.

This commit makes sure the subscriber's interests are created in a
symmetrical way to the publisher side which solves the discovery
problem.

Also when creating the publisher's discovery interest use the wildcard
option. This is primarily so that discovery information is sent out
on localhost given the above change.

Signed-off-by: Andriy Gelman <andriy.gelman@gmail.com>